+++ src/app/views/discover/games/view/followers/followers.route.ts
@@ -3,5 +3,5 @@ import { RouteRecordRaw } from 'vue-router';
 export const routeDiscoverGamesViewFollowers: RouteRecordRaw = {
 	name: 'discover.games.view.followers',
 	path: 'followers',
-	component: () => import('./followers.vue'),
+	component: () => import('./RouteProfileFollowers.vue'),
 };
